requirements.txt
rdkit-pypi>=2023.9.1
meeko>=0.7.0
scipy>=1.10.0
gemmi>=0.6.0
numpy>=1.24.0
